refactor: 优化代码
