+++ src/Plugin/Wechat/Ecommerce/Refund/ApplyPlugin.php
@@ -9,7 +9,7 @@ use Yansongda\Pay\Exception\InvalidParamsException;
 use Yansongda\Pay\Plugin\Wechat\GeneralPlugin;
 use Yansongda\Pay\Rocket;
 
-class FindReturnAdvancePlugin extends GeneralPlugin
+class ApplyPlugin extends GeneralPlugin
 {
     /**
      * @throws \Yansongda\Pay\Exception\InvalidParamsException
@@ -19,32 +19,30 @@ class FindReturnAdvancePlugin extends GeneralPlugin
         throw new InvalidParamsException(Exception::NOT_IN_SERVICE_MODE);
     }
 
+    protected function getPartnerUri(Rocket $rocket): string
+    {
+        return 'v3/ecommerce/refunds/apply';
+    }
+
     /**
      * @throws \Yansongda\Pay\Exception\ContainerDependencyException
      * @throws \Yansongda\Pay\Exception\ContainerException
-     * @throws \Yansongda\Pay\Exception\InvalidParamsException
      * @throws \Yansongda\Pay\Exception\ServiceNotFoundException
      */
-    protected function getPartnerUri(Rocket $rocket): string
-    {
-        $payload = $rocket->getPayload();
-        $config = get_wechat_config($rocket->getParams());
-        $subMchId = $payload->get('sub_mchid', $config->get('sub_mch_id', ''));
-
-        if (is_null($payload->get('refund_id'))) {
-            throw new InvalidParamsException(Exception::MISSING_NECESSARY_PARAMS);
-        }
-
-        return 'v3/ecommerce/refunds/'.$payload->get('refund_id').'/return-advance?sub_mchid='. $subMchId;
-    }
-
-    protected function getMethod(): string
-    {
-        return 'GET';
-    }
-
     protected function doSomething(Rocket $rocket): void
     {
-        $rocket->setPayload(null);
+        $config = get_wechat_config($rocket->getParams());
+        $payload = $rocket->getPayload();
+
+        $wechatId = [
+            'sub_mchid' => $payload->get('sub_mchid', $config->get('sub_mch_id', '')),
+            'sp_appid' => $config->get('mp_app_id', ''),
+        ];
+
+        if (!$payload->has('notify_url')) {
+            $wechatId['notify_url'] = $config->get('notify_url');
+        }
+
+        $rocket->mergePayload($wechatId);
     }
 }
