+++ src/Plugin/Wechat/Ecommerce/Refund/FindPlugin.php
@@ -9,7 +9,7 @@ use Yansongda\Pay\Exception\InvalidParamsException;
 use Yansongda\Pay\Plugin\Wechat\GeneralPlugin;
 use Yansongda\Pay\Rocket;
 
-class FindReturnAdvancePlugin extends GeneralPlugin
+class FindPlugin extends GeneralPlugin
 {
     /**
      * @throws \Yansongda\Pay\Exception\InvalidParamsException
@@ -31,11 +31,15 @@ class FindReturnAdvancePlugin extends GeneralPlugin
         $config = get_wechat_config($rocket->getParams());
         $subMchId = $payload->get('sub_mchid', $config->get('sub_mch_id', ''));
 
-        if (is_null($payload->get('refund_id'))) {
-            throw new InvalidParamsException(Exception::MISSING_NECESSARY_PARAMS);
+        if (!is_null($payload->get('refund_id'))) {
+            return 'v3/ecommerce/refunds/id/'.$payload->get('refund_id').'?sub_mchid='. $subMchId;
         }
 
-        return 'v3/ecommerce/refunds/'.$payload->get('refund_id').'/return-advance?sub_mchid='. $subMchId;
+        if (!is_null($payload->get('out_refund_no'))) {
+            return 'v3/ecommerce/refunds/out-refund-no/'.$payload->get('out_refund_no').'?sub_mchid='. $subMchId;
+        }
+
+        throw new InvalidParamsException(Exception::MISSING_NECESSARY_PARAMS);
     }
 
     protected function getMethod(): string
